+++ src/Utility/LinkScanner.php
@@ -26,6 +26,7 @@ use Exception;
 use InvalidArgumentException;
 use LinkScanner\ResultScan;
 use LinkScanner\ScanEntity;
+use PHPUnit\Framework\Exception as PHPUnitException;
 use RuntimeException;
 use Serializable;
 use Tools\BodyParser;
@@ -178,7 +179,7 @@ class LinkScanner implements Serializable
             $this->lockFile
         ), RuntimeException::class);
 
-        return $this->getConfig('lockFile') ? (new Filesystem())->createFile($this->lockFile) : true;
+        return $this->getConfig('lockFile') ? Filesystem::instance()->createFile($this->lockFile) : true;
     }
 
     /**
@@ -189,7 +190,7 @@ class LinkScanner implements Serializable
      */
     protected function _getAbsolutePath(string $filename): string
     {
-        return (new Filesystem())->makePathAbsolute($filename, $this->getConfig('target'));
+        return Filesystem::instance()->makePathAbsolute($filename, $this->getConfig('target'));
     }
 
     /**
@@ -227,6 +228,8 @@ class LinkScanner implements Serializable
             if ($this->getConfig('cache') && ($response->isOk() || $response->isRedirect())) {
                 Cache::write($cacheKey, [$response, (string)$response->getBody()], 'LinkScanner');
             }
+        } catch (PHPUnitException $e) {
+            throw $e;
         } catch (Exception $e) {
             $response = (new Response())->withStatus(404);
         }
@@ -408,7 +411,7 @@ class LinkScanner implements Serializable
         }
 
         $filename = $this->_getAbsolutePath($filename ?: sprintf('results_%s_%s', $this->hostname, $this->startTime));
-        (new Filesystem())->createFile($filename, serialize($this));
+        Filesystem::instance()->createFile($filename, serialize($this));
         $this->dispatchEvent('LinkScanner.resultsExported', [$filename]);
 
         return $filename;