+++ tests/test_app/TestApp/config/routes.php
@@ -7,7 +7,8 @@ use Cake\Routing\Router;
 
 Router::defaultRouteClass(DashedRoute::class);
 
-Router::scope('/', function (RouteBuilder $routes) {
+/** @var \Cake\Routing\RouteBuilder $routes */
+$routes->scope('/', function (RouteBuilder $routes) {
     $routes->connect('/', ['controller' => 'Pages', 'action' => 'display', 'home']);
     $routes->redirect('/pages/redirect', ['controller' => 'Pages', 'action' => 'display', 'third_page'], ['status' => 302]);
     $routes->redirect('/pages/sameredirect', ['controller' => 'Pages', 'action' => 'display', 'third_page'], ['status' => 302]);